+++ Sources/Common/Background Queue/Type/QueueTaskType.swift
@@ -1,5 +1,10 @@
 import Foundation
 
-enum QueueTaskType: String {
+public enum QueueTaskType: String {
     case trackDeliveryMetric
+    case identifyProfile
+    case trackEvent
+    case registerPushToken
+    case deletePushToken
+    case trackPushMetric
 }