+++ Sources/Common/Extensions/UIKitExtensions.swift
@@ -10,4 +10,10 @@ public extension UIApplication {
     }
 }
 
+public extension UIViewController {
+    // find the bundleId of the framework this View belongs in. This can be used to differentiate Views that belong to host app and Views that belong to SDKs/frameworks.
+    var bundleIdOfView: String? {
+        Bundle(for: type(of: self)).bundleIdentifier
+    }
+}
 #endif