+++ Sources/Tracking/Service/Request/RegisterDeviceRequest.swift
@@ -1,6 +1,10 @@
 import Foundation
 
 // https://customer.io/docs/api/#operation/add_device
-struct RegisterDeviceRequest<T: Encodable>: Encodable {
-    let device: Device<T>
+public struct RegisterDeviceRequest<T: Encodable>: Encodable {
+    public let device: Device<T>
+
+    public init(device: Device<T>) {
+        self.device = device
+    }
 }